Creates Multi Agent DDPG. One model controlling the blues.

+++ envs/agents/deepvss/lib/common.py
@@ -12,7 +12,7 @@ import torch
 def unpack_batch(batch):
     states, actions, rewards, dones, last_states = [], [], [], [], []
     for exp in batch:
-        state = np.array(exp.state, copy=False)
+        state = np.array(exp.state, copy=False, dtype=np.float)
         states.append(state)
         actions.append(exp.action)
         rewards.append(exp.reward)
@@ -20,7 +20,7 @@ def unpack_batch(batch):
         if exp.last_state is None:
             last_states.append(state)  # the result will be masked anyway
         else:
-            last_states.append(np.array(exp.last_state, copy=False))
+            last_states.append(np.array(exp.last_state, copy=False, dtype=np.float))
     return np.array(states, copy=False), np.array(actions), np.array(rewards, dtype=np.float32), \
         np.array(dones, dtype=np.uint8), np.array(last_states, copy=False)
